Organic_Website-Angular-/Backend/requirement.txt.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cssselect2==0.7.0
Cython==3.0.0
defusedxml==0.7.1
distro==1.9.0
Django==3.2.19
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
exceptiongroup==1.2.2
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.8
inflection==0.5.1
jiter==0.5.0
multidict==6.1.0
mysqlclient==2.2.0
numpy==2.0.2
odfpy==1.4.1
openai==0.28.0
packaging==23.1
pdfkit==1.0.0
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydyf==0.11.0
PyJWT==2.8.0
pyphen==0.16.0
pytz==2023.3
PyYAML==5.4.1
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tinycss2==1.3.0
tokenizers==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
webencodings==0.5.1
yarl==1.11.1
zopfli==0.2.3